+++ bcs-app/backend/bcs_web/verfiy/constants.py
@@ -1,0 +1,40 @@
+# -*- coding: utf-8 -*-
+#
+# Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) available.
+# Copyright (C) 2017-2019 THL A29 Limited, a Tencent company. All rights reserved.
+# Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://opensource.org/licenses/MIT
+#
+# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+# an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+#
+import logging
+
+from backend.utils.basic import ChoicesEnum
+
+logger = logging.getLogger(__name__)
+
+
+class PermMultiOperator(ChoicesEnum):
+    # 交集
+    AND = 'and'
+    # 或集
+    OR = 'or'
+
+    _choices_labels = (
+        (AND, 'AND'),
+        (OR, 'OR'),
+    )
+
+
+# verify resource code for perm
+verify_resource_exist = False
+
+
+try:
+    from .constants_ext import verify_resource_exist
+except ImportError as e:
+    logger.debug('Load extension failed: %s', e)
