+++ bcs-app/backend/bcs_web/verfiy/views.py
@@ -17,13 +17,14 @@ from django.utils.translation import ugettext_lazy as _
 from rest_framework import viewsets
 
 from backend.accounts import bcs_perm
-from backend.apps.verfy import constants, serializers
 from backend.components import paas_cc
 from backend.utils import FancyDict
 from backend.utils.errcodes import ErrorCode
 from backend.utils.error_codes import error_codes
 from backend.utils.exceptions import APIError, NoAuthPermError, VerifyAuthPermError, VerifyAuthPermErrorWithNoRaise
 from backend.utils.response import APIResult
+
+from . import constants, serializers
 
 logger = logging.getLogger(__name__)
