+++ bcs-app/backend/settings/base.py
@@ -71,7 +71,7 @@ INSTALLED_APPS = [
     "backend.apps.ticket",
     "backend.helm.app",
     "backend.helm.helm",
-    "backend.apps.hosts",
+    "backend.container_service.infras.hosts.terraform",
     # 模板集功能模块
     "backend.templatesets.var_mgmt.apps.VarMgmtConfig",
 ]
