+++ bcs-app/backend/urls.py
@@ -54,7 +54,7 @@ urlpatterns = [
     url(r"^", include("backend.uniapps.application.urls")),
     url(r"^", include("backend.activity_log.urls")),
     # 权限验证
-    url(r"^", include("backend.apps.verfy.urls")),
+    url(r"^", include("backend.apps.verify.urls")),
     url(r"^api-auth/", include("rest_framework.urls")),
     # BCS K8S special urls
     url(r"^", include("backend.helm.helm.urls")),
